// exercise/src/models/RootStackParams.ts
type RootStackParams = {
  Home: { typology: string };
  Detail: {id?: string | undefined;
          name?: string | undefined;}
  HomeStack: undefined;
  Profile: undefined;
  Favorite:  undefined
  Homepage: undefined;
  SignUp: undefined;
  Login: undefined;
  Editor: undefined;
  Personal: undefined;
};

export default RootStackParams;